+++ src/Storefront/Controller/ResultController.php
@@ -58,10 +58,30 @@ class ResultController extends StorefrontController
         );
         $response->setContent(
             $recordList->getContent(
-                (string) parse_url($_SERVER['REQUEST_URI'] ?? '', PHP_URL_QUERY)
+                $this->parseQueryString((string) parse_url($_SERVER['REQUEST_URI'] ?? '', PHP_URL_QUERY))
             )
         );
 
         return $response;
+    }
+
+    private function parseQueryString(string $queryString): string
+    {
+        if ('' === $queryString) {
+            return '';
+        }
+
+        $queryParams = explode('&', $queryString);
+        $result = array_reduce(
+            $queryParams,
+            function (string $carry, string $queryParam) {
+                $result = explode('=', $queryParam);
+
+                return $carry . sprintf('&%s=%s', $result[0], htmlspecialchars($result[1]));
+            },
+            ''
+        );
+
+        return substr($result, 1);
     }
 }